refactor: memo를 사용해 불필요한 리렌더링 제거
